Restored cross button

+++ app/admin/tracks.rb
@@ -177,7 +177,7 @@ ActiveAdmin.register Track do
       f.input :file, as: :file , label: "Music File"
       div class: 'file-hint' do
         span 'Existing File: ' + file_hint(f.object), id: 'hint'
-        span link_to 'x', remove_file_admin_track_path(f.object), class: 'remove-file', method: :delete, remote: true if f.object.file.blob&.persisted?
+        span link_to 'x', remove_file_admin_track_path(f.object), class: 'remove-file', data: { confirm: 'Are you sure you want to remove this audio?' }, method: :delete, remote: true if f.object.file.blob&.persisted?
       end
       f.input :description, input_html: { class: 'autogrow', rows: 4, cols: 20 }
       f.input :status, as: :select, collection: tracks_status_list, include_blank: false
